Calcular Volumen

+++ encuesta/forms.py
@@ -1,8 +1,8 @@
 from django import forms
 
 class OperacionesForm(forms.Form):
-    num1 = forms.IntegerField(label="Número 1")
-    num2 = forms.IntegerField(label="Número 2")
+    num1 = forms.FloatField(label="Número 1")
+    num2 = forms.FloatField(label="Número 2")
     operaciones = [
         ("suma", "+"),
         ("resta", "-"),
